Added ability to assign fit_pdfs_batch.sh only a portion of a galstar output tarball.

+++ scripts/fit_pdfs_batch.sh
@@ -9,6 +9,8 @@
 #    TARIN      = filename of tarball with galstar outputs
 #    OUTFN      = file to output DM,Ar relation to for each pixel
 #    ARGS       = optional arguments to pass to fit_pdfs.py
+#    DIV        = # of processes working on this output file
+#    PART       = part # of file this process is to complete
 #
 
 # Determine the temporary directory and working directory
@@ -37,24 +39,50 @@ cd $tmpdir
 
 # Run fit_pdfs.py on each pixel
 counter=1
+thispart=1
+partindex=1
+partsize=`expr $npix / $DIV`
+remainder=`expr $npix % $DIV`
+remainderused=0
 for binfile in $binfilelist; do
-	# Determine filenames corresponding to current pixel
-	pixindex=${binfile%_DM_Ar.dat}
-	statsfn="$pixindex.stats"
-	binfn="${pixindex}_DM_Ar.dat"
+	# Determine whether to process this pixel
+	if [ "$thispart" -gt "$partsize" ]; then
+		if [ "$remainder" -gt "0" ]; then
+			if [ "$remainderused" -eq "0" ]; then
+				remainderused=1
+				remainder=`expr $remainder - 1`
+			else
+				partindex=`expr $partindex + 1`
+				thispart=2
+				remainderused=0
+			fi
+		else
+			partindex=`expr $partindex + 1`
+		fi
+	else
+		thispart=`expr $thispart + 1`
+	fi
 	
-	# Extract current pixel
-	tar -xf $tarin $statsfn $binfn
-	echo "$counter of $npix: Fitting l.o.s. reddening law for healpix pixel $pixindex ..."
-	
-	# Run fit_pdfs.py on this pixel
-	$scriptsdir/fit_pdfs.py $binfn $statsfn -o $outfn $pixindex $ARGS
-	
-	# Remove temporary files
-	rm $statsfn $binfn
+	if [ $partindex -eq $PART ]; then
+		# Determine filenames corresponding to current pixel
+		pixindex=${binfile%_DM_Ar.dat}
+		statsfn="$pixindex.stats"
+		binfn="${pixindex}_DM_Ar.dat"
+		
+		# Extract current pixel
+		tar -xf $tarin $statsfn $binfn
+		echo "$counter of $npix: Fitting l.o.s. reddening law for healpix pixel $pixindex ..."
+		
+		# Run fit_pdfs.py on this pixel
+		$scriptsdir/fit_pdfs.py $binfn $statsfn -o $outfn $pixindex $ARGS
+		
+		# Remove temporary files
+		rm $statsfn $binfn
+		
+		echo ""
+	fi
 	
 	counter=`expr $counter + 1`
-	echo ""
 done
 
 gzip -9 $outfn
